UIBULKED-117 refactoring
